% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-plot.R
\name{adapt_1d_plot}
\alias{adapt_1d_plot}
\title{Plotting}
\usage{
adapt_1d_plot(obj, alpha, title, data = NULL, xlab = "x", disp_pmax = 0.2,
  disp_lfdrmax = 0.2, num_xbreaks = 3, rand_seed_perturb = NA,
  legend_pos = "topright", ...)
}
\description{
Plotting
}
