% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-wrappers.R
\name{adapt_gam}
\alias{adapt_gam}
\title{Adaptive P-value Thresholding with Generalized Additive Models}
\usage{
adapt_gam(x, pvals, pi_formulas, mu_formulas, piargs = list(),
  muargs = list(), dist = beta_family(), s0 = rep(0.45, length(pvals)),
  alphas = seq(0.01, 1, 0.01), ...)
}
\arguments{
\item{x}{a data.frame. Covariates (i.e. side-information)}

\item{pvals}{a vector of values in [0, 1]. P-values}

\item{pi_formulas}{a vector/list of strings/formulas. Formulas for fitting pi(x) by gam. See Details}

\item{mu_formulas}{a vector/list of strings/formulas. Formulas for fitting mu(x) by gam. See Details}

\item{piargs}{a list. Other arguments passed to gam for fitting pi(x)}

\item{muargs}{a list. Other arguments passed to gam for fitting mu(x)}

\item{dist}{an object of class "\code{\link{exp_family}}". \code{\link{beta_family}()} as default}

\item{s0}{a vector of values in [0, 0.5). Initial threshold.}

\item{alphas}{a vector of values in (0, 1). Target FDR levels.}

\item{...}{other arguments passed to \code{\link{adapt}}}
}
\description{
Adaptive P-value Thresholding with Generalized Additive Models
}
