% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-wrappers.R
\name{adapt_glm}
\alias{adapt_glm}
\title{Easy-to-use Wrappers of Adaptive P-value Thresholding}
\usage{
adapt_glm(x, pvals, pi_formulas, mu_formulas, dist = beta_family(),
  s0 = rep(0.45, length(pvals)), alphas = seq(0.01, 1, 0.01), ...)
}
\description{
Easy-to-use Wrappers of Adaptive P-value Thresholding
}
