% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-wrappers.R
\name{adapt_glmnet}
\alias{adapt_glmnet}
\title{Adaptive P-value Thresholding with L1/L2 Penalized Generalized Linear Models}
\usage{
adapt_glmnet(x, pvals, piargs = list(), muargs = list(),
  dist = beta_family(), s0 = rep(0.45, length(pvals)), alphas = seq(0.01,
  1, 0.01), ...)
}
\arguments{
\item{x}{a data.frame. Covariates (i.e. side-information)}

\item{pvals}{a vector of values in [0, 1]. P-values}

\item{piargs}{a list. Other arguments passed to glmnet for fitting pi(x)}

\item{muargs}{a list. Other arguments passed to glmnet for fitting mu(x)}

\item{dist}{an object of class "\code{\link{exp_family}}". \code{\link{beta_family}()} as default}

\item{s0}{a vector of values in [0, 0.5). Initial threshold.}

\item{alphas}{a vector of values in (0, 1). Target FDR levels.}

\item{...}{other arguments passed to \code{\link{adapt}}}
}
\description{
Adaptive P-value Thresholding with L1/L2 Penalized Generalized Linear Models
}
