% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep_mix.R
\name{Estep_mix}
\alias{Estep_mix}
\title{Computing E-step for Mixture Models}
\usage{
Estep_mix(pvals, s, dist, pix, mux)
}
\arguments{
\item{pvals}{a vector of values in [0, 1]. P-values}

\item{s}{a vector of values in [0, 1]. Threshold curve}

\item{dist}{an object of class "\code{\link{exp_family}}"}

\item{pix}{a vector of values in [0, 1]. \eqn{\pi(x_{i})}{\pi(xi)}}

\item{mux}{a vector of values. \eqn{\mu(x_{i})}{\mu(xi)}}
}
\value{
Imputed values, including
\item{Hhat}{a vector of values in [0, 1]. Imputed values for \eqn{H_{i}}{Hi}'s}
\item{phat}{a vector of values in [0, 1]. Imputed values for \eqn{p_{i}}{pi}'s}
}
\description{
\code{Estep_mix} computes the E-step, namely imputation of missing values, including the p-values and indicators of null/non-null hypotheses.
}
\details{
The p-values are assumed to be generated from an covariate-varying mixture model
\deqn{H_{i}\sim Ber(\pi(x_{i})), p_{i} \mid (H_{i} = 0) \sim U([0, 1]), p_{i} \mid (H_{i} = 1) \sim h(\cdot; \mu(x_{i}))}{Hi ~ Ber(\pi(xi)), pi | (Hi = 0) ~ U([0, 1]), pi | (Hi = 1) ~ h(p; \mu(xi))}
where \eqn{h(p; \mu)} is the density of an exponential family (see \code{\link{exp_family}}). Given a threshold curve s(x), the partially i-th masked p-value is defined as \eqn{p_{i}}{pi} if \eqn{p_{i}\in [s(x_{i}), 1-s(x_{i})]}{pi in [s(xi), 1-s(xi)]}. The E-step computes the expectation of \eqn{H_{i}}{Hi} and \eqn{g(p_{i})H_{i}}{g(pi)Hi} given the parameters \eqn{\pi(x_{i})}{\pi(xi)} and \eqn{\mu(x_{i})}{\mu(xi)}, based on which computes the imputed values for \eqn{H_{i}}{Hi} and \eqn{p_{i}}{pi}. See ...
}
\keyword{internal}
