% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estrogen}
\alias{estrogen}
\title{Gene/Drug response dataset}
\format{An object of class \code{data.frame} with 22283 rows and 3 columns.}
\usage{
estrogen
}
\description{
Microarray data of 22283 genes on breast cancer cells in response to estrogen
}
\keyword{datasets}
