% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_family.R
\name{exp_family}
\alias{exp_family}
\alias{beta_family}
\alias{inv_gaussian_family}
\title{An S3generic to define an exponential family}
\usage{
exp_family(g, ginv, eta, mustar, A, name = NULL, family = NULL)

beta_family()

inv_gaussian_family()
}
\arguments{
\item{g}{function. An transformation of p-values}

\item{ginv}{function. The inverse function of \code{g}}

\item{eta}{function. The natural parameter as a function of the mean parameter \code{mu}}

\item{mustar}{scalar. The mean parameter that gives \eqn{U([0, 1])}}

\item{A}{function. The partition function}

\item{name}{character. A name for the family. NULL by default}

\item{family}{object of S3 class "\code{\link[stats]{gaussian}}" from \code{stats} package. The family used for model fitting in \code{glm}, \code{gam}, \code{glmnet}, etc..}
}
\value{
an object of S3 class "exp_family". This includes all inputs and  \item{h }{function. The density function computed}
}
\description{
An object containing all required information in an exponential
family to perform the E-step. The exponential function is encoded by
\deqn{h(p; \eta) = \exp\{(\eta(\mu) - \eta(\mu^{*})) g(p) - (A(\mu) - A(\mu^{*}))\}}{h(p; \eta) = exp{(\eta(\mu) - \eta(\mu*)) g(p) - (A(\mu) - A(\mu*))}}
where \eqn{g(p)} is an arbitrary transformation, \eqn{\mu} is the
\emph{mean parameter}, \eqn{\eta} is the natural parameter,
and \eqn{A(\mu)} is the partition function. The extra redundant
parameter \eqn{\mu^{*}}{\mu*} is to guarantee that \eqn{U([0, 1])}
belongs to the class.
}
\seealso{
\code{\link{prod}} for products, \code{\link{cumsum}} for cumulative
   sums, and \code{\link{colSums}}/\code{\link{rowSums}} marginal sums over
   high-dimensional arrays.
}
