% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fdp.hat}
\alias{fdp.hat}
\title{Function to calculate FDPhat}
\usage{
fdp.hat(pvals, s)
}
\arguments{
\item{pvals:}{p-values}

\item{s:}{threshold curve}
}
\value{
fdphat: estimated FDP
}
\description{
Function to calculate FDPhat
}
