% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-plot.R
\name{plot_1d}
\alias{plot_1d}
\alias{plot_1d_thresh}
\alias{plot_1d_lfdr}
\alias{plot_1d_params}
\title{Plotting for 1d Covariate}
\usage{
plot_1d_thresh(obj, alpha, title, data = NULL, xlab = "x", xlim = NULL,
  disp_ymax = 0.2, num_yticks = 3, rand_seed_perturb = NA, ...)

plot_1d_lfdr(obj, alpha, title, data = NULL, xlab = "x", xlim = NULL,
  disp_ymax = 0.2, num_yticks = 3, legend_pos = "topright", ...)

plot_1d_params(obj, alpha, title, data = NULL, xlab = "x", xlim = NULL,
  num_yticks = 3, ...)
}
\arguments{
\item{obj}{an 'adapt' object}

\item{alpha}{a positive scalar in (0, 1). Target FDR level}

\item{title}{a string. Title of the figure}

\item{data}{a list in the form of list(x = , pvals = ). NULL if obj$data is not NULL}

\item{xlab}{a string. Label of the x-axis}

\item{xlim}{a vector of length 2. Limits of x-axis}

\item{disp_ymax}{a positive scalar in (0, 1]. Maximum value displayed in the y-axis}

\item{num_yticks}{a positive integer. Number of ticks in the y-axis}

\item{rand_seed_perturb}{random seed if jitter is added. NA if no jittering is needed}

\item{...}{other arguments passed to \code{\link[graphics]{par}}}

\item{legend_pos}{a string. Position of the legend}
}
\description{
Plotting of the outputs of \code{adapt} when \code{x} is 1-dimensional, including threshold curves, level curves of local FDR and parameter estimates.
}
