% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-plot-2d.R
\name{plot_2d}
\alias{plot_2d}
\alias{plot_2d_thresh}
\alias{plot_2d_lfdr}
\title{Plotting Functions for AdaPT with 2D Covariates}
\usage{
plot_2d_thresh(obj, alpha, title, data = NULL, xlab = NULL, ylab = NULL,
  keyaxes = list(), ...)

plot_2d_lfdr(obj, alpha, title, targetp, data = NULL, xlab = NULL,
  ylab = NULL, keyaxes = list(), ...)
}
\arguments{
\item{obj}{an 'adapt' object}

\item{alpha}{a positive scalar in (0, 1). Target FDR level}

\item{title}{a string. Title of the figure}

\item{data}{a list in the form of list(x = , pvals = ). NULL if obj$data is not NULL}

\item{xlab, ylab}{a string. Label of x/y-axis}

\item{keyaxes}{a list of arguments passed into axis. The graphical setting for the legend bar. An empty list by default}

\item{...}{other arguments passed to \code{\link[graphics]{par}}}

\item{targetp}{a real in (0, 1). See Details}
}
\description{
Plotting the outputs of \code{adapt} when \code{x} is 2-dimensional, including threshold curves and level curves of local FDR.
}
\details{
The breaks in the legend of \code{plot_2d_thresh} correspond to the maximum, the 95% quantile, the 90% quantile, the 85% quantile, the 75% quantile and the minimum of the thresholds.

\code{plot_2d_lfdr} gives the contour plot of local FDR estimates when all p-values are equal to \code{targetp}. It is recommended to run \code{plot_2d_lfdr} for multiple \code{targetp}'s ranging from {0.001, 0.005, 0.01, 0.05}.
}
