% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-plot.R
\name{plot_thresh_1d}
\alias{plot_thresh_1d}
\title{Plotting}
\usage{
plot_thresh_1d(obj, alpha, title, data = NULL, xlab = "x",
  disp_ymax = 0.2, num_xbreaks = 3, rand_seed_perturb = NA, ...)
}
\description{
Plotting
}
