% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_model.R
\name{safe_fit}
\alias{safe_fit}
\title{Securing model-fitting algorithm (internal function)}
\usage{
safe_fit(algo, data_args, algo_args, alter_args = NULL, ...)
}
\arguments{
\item{algo}{A function (fitting algorithm).}

\item{data_args}{A list, with the keys being the name of an argument of \code{algo} for specifying
data/input. Examples include list(data = ...) or list(x = ..., y = ...).}

\item{algo_args}{A list, with the keys being the name of an argument of \code{algo}, other than
those specified for data/input, and the values being the value of that argument. Examples
include list(formula = ...).}

\item{alter_args}{A list of list, with each sub-list of the same lengths and keys as
\code{algo_args}, but with different values. NULL if no candidate argument.}
}
\value{
One of the following: \itemize{
   \item The output of \code{algo}, when running \code{algo} on \code{algo_args} does not report an error,
   \item an error message indicating the error by running \code{algo} on \code{algo_args}, and the output
   of \code{algo} on the first entry of \code{alter_args} which does not report an error,
   \item an error message indicating the error by running \code{algo} on \code{algo_args}, together with
   an error message indicating that none of candidate arguments work.
   }
}
\description{
A generic handle for making a fitting algorithm secure to run by adding a list of
candidate arguments. When the fitting function encounters an error, it prints the
current error message and proceeds to run the function on the candidate argument,
one-by-one from the list. It stops when a particular set of arguments does not give
an error and the results of which will be outputed. If all sets of candidate arguments
report errors, the function will report an error message, in addition to the previous
one.
}
